move rendering to the type engine and fix some issues
