+++ modules/module-data-browser/src/core/query/widgets/CellDisplayWidget.tsx
@@ -1,12 +1,4 @@
-import {
-  CheckboxWidget,
-  ImageMedia,
-  ioc,
-  MetadataWidget,
-  SmartDateDisplayWidget,
-  styled
-} from '@journeyapps-labs/reactor-mod';
-import { Attachment, Day, Location } from '@journeyapps/db';
+import { ioc, SmartDateDisplayWidget, styled } from '@journeyapps-labs/reactor-mod';
 import * as _ from 'lodash';
 import * as React from 'react';
 import { PageRow } from '../Page';
@@ -15,12 +7,6 @@ import { TypeEngine } from '../../../forms/TypeEngine';
 namespace S {
   export const Empty = styled.div`
     opacity: 0.2;
-  `;
-
-  export const Preview = styled.img`
-    max-height: 40px;
-    max-width: 40px;
-    cursor: pointer;
   `;
 
   export const pill = styled.div`
@@ -34,14 +20,6 @@ namespace S {
     display: flex;
     column-gap: 2px;
     row-gap: 2px;
-  `;
-
-  export const Max = styled.div`
-    max-width: 500px;
-    white-space: pre;
-    display: inline;
-    overflow: hidden;
-    text-overflow: ellipsis;
   `;
 }
 
@@ -58,15 +36,22 @@ export const CellDisplayWidget: React.FC<CellDisplayWidgetProps> = (props) => {
   if (cell == null) {
     return <S.Empty>null</S.Empty>;
   }
-  if (_.isString(cell)) {
-    if (cell.trim() === '') {
-      return <S.Empty>empty</S.Empty>;
-    }
-    return <S.Max>{cell}</S.Max>;
+
+  if (name === 'updated_at') {
+    return <SmartDateDisplayWidget date={cell} />;
   }
-  if (_.isNumber(cell)) {
-    return cell;
+
+  let display = ioc.get(TypeEngine).getHandler(row.definition.definition.attributes[name].type)?.generateDisplay({
+    model: row.model,
+    value: cell,
+    label: row.definition.definition.attributes[name].label,
+    name,
+    type: row.definition.definition.attributes[name].type
+  });
+  if (display) {
+    return display;
   }
+
   if (_.isArray(cell)) {
     if (cell.length === 0) {
       return <S.Empty>empty array</S.Empty>;
@@ -81,53 +66,7 @@ export const CellDisplayWidget: React.FC<CellDisplayWidgetProps> = (props) => {
       </S.Pills>
     );
   }
-  if (cell instanceof Date) {
-    return <SmartDateDisplayWidget date={cell} />;
-  }
-  if (cell instanceof Day) {
-    return <SmartDateDisplayWidget date={cell.toDate()} />;
-  }
-  if (_.isBoolean(cell)) {
-    return (
-      <CheckboxWidget
-        checked={cell}
-        onChange={(checked) => {
-          row.model.set(name, checked);
-        }}
-      />
-    );
-  }
-  if (cell instanceof Location) {
-    return (
-      <>
-        <MetadataWidget label={'Lat'} value={`${cell.latitude}`} />
-        <MetadataWidget label={'Long'} value={`${cell.longitude}`} />
-      </>
-    );
-  }
-  if (cell instanceof Attachment) {
-    if (cell.uploaded()) {
-      return (
-        <S.Preview
-          onClick={() => {
-            ioc
-              .get(TypeEngine)
-              .getHandler(row.definition.definition.attributes[name].type)
-              .decode(cell)
-              .then((media: ImageMedia) => {
-                if (media instanceof ImageMedia) {
-                  media.open();
-                } else {
-                  window.open(cell.url(), '_blank');
-                }
-              });
-          }}
-          src={cell.urls['thumbnail']}
-        />
-      );
-    }
-    return <S.Empty>Not uploaded</S.Empty>;
-  }
+
   console.log('unknown type', cell);
   return null;
 };
